Fix bugs, refine naming and types, add examples
Actually a tonne of changes I kind of kept thinking of 'little' things
I wanted to do and it ended up being a lot.
Mostly reshuffled a bunch of stuff, changed types and names to be simpler.
Allow for Syntax to pass context down so Decode.reference can make allowances when it is referencing in same file.
Add some examples.
Fix string literal writing.